Renamed all BlogXXX related components to EditorialXXX, removed margins for these component containers as these should be in the projects using them, deleted Publication component as it's just an implementation
