--- packages/asc-ui/src/components/Blog/BlogBody/BlogBody.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import BlogBodyStyle, { Props } from './BlogBodyStyle'
-
-const BlogBody: React.FC<Props & React.HTMLAttributes<HTMLElement>> = ({
-  children,
-  ...otherProps
-}) => <BlogBodyStyle {...otherProps}>{children}</BlogBodyStyle>
-
-export default BlogBody
