--- packages/asc-ui/src/components/Blog/BlogBody/index.ts
@@ -1,2 +1,0 @@
-export { default as BlogBodyStyle } from './BlogBodyStyle'
-export { default as BlogBody } from './BlogBody'
