--- packages/asc-ui/src/components/Blog/BlogContent/BlogContent.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import BlogContentStyle from './BlogContentStyle'
-
-const BlogContent: React.FC<React.HTMLAttributes<HTMLElement>> = ({
-  children,
-  ...otherProps
-}) => <BlogContentStyle {...otherProps}>{children}</BlogContentStyle>
-
-export default BlogContent
