--- packages/asc-ui/src/components/Blog/BlogContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BlogContent'
