--- packages/asc-ui/src/components/Blog/BlogHeader/BlogHeader.tsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import BlogHeaderStyle from './BlogHeaderStyle'
-import Heading from '../../Heading/Heading'
-
-export type Props = {
-  title: string
-} & React.HTMLAttributes<HTMLElement>
-
-const BlogHeader: React.FC<Props> = ({ title, children, ...otherProps }) => (
-  <BlogHeaderStyle {...otherProps}>
-    <Heading as="h1">{title}</Heading>
-    {children}
-  </BlogHeaderStyle>
-)
-
-export default BlogHeader
