--- packages/asc-ui/src/components/Blog/BlogHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BlogHeader'
