--- packages/asc-ui/src/components/Blog/BlogMetaList/index.ts
@@ -1,8 +1,0 @@
-import BlogMetaListStyle, { BlogMetaListItemStyle } from './BlogMetaListStyle'
-
-export { default } from './BlogMetaList'
-
-export const BlogMetaListItemStyles = {
-  BlogMetaListStyle,
-  BlogMetaListItemStyle,
-}
