--- packages/asc-ui/src/components/Blog/BlogPost/index.ts
@@ -1,2 +1,0 @@
-export { default as BlogPostStyle } from './BlogPostStyle'
-export { default as BlogPost } from './BlogPost'
