--- packages/asc-ui/src/components/Blog/BlogSidebar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BlogSidebar'
