--- packages/asc-ui/src/components/Blog/index.ts
@@ -1,20 +1,0 @@
-import BlogContentStyle from './BlogContent/BlogContentStyle'
-import BlogHeaderStyle from './BlogHeader/BlogHeaderStyle'
-import BlogSidebarStyle from './BlogSidebar/BlogSidebarStyle'
-import BlogMetaListStyle from './BlogMetaList/BlogMetaListStyle'
-import { BlogPostStyle } from './BlogPost'
-
-export { BlogPost } from './BlogPost'
-export { BlogBody } from './BlogBody'
-export { default as BlogContent } from './BlogContent'
-export { default as BlogSidebar } from './BlogSidebar'
-export { default as BlogMetaList } from './BlogMetaList'
-export { default as BlogHeader } from './BlogHeader'
-
-export const BlogStyles = {
-  BlogPostStyle,
-  BlogContentStyle,
-  BlogMetaListStyle,
-  BlogSidebarStyle,
-  BlogHeaderStyle,
-}
