+++ packages/asc-ui/src/components/Editorial/Article/article_feed_example.json
