+++ packages/asc-ui/src/components/Editorial/EditorialBody/index.ts
@@ -1,0 +1,2 @@
+export { default as EditorialBodyStyle } from './EditorialBodyStyle'
+export { default as EditorialBody } from './EditorialBody'
