+++ packages/asc-ui/src/components/Editorial/EditorialContent/EditorialContent.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import EditorialContentStyle from './EditorialContentStyle'
+
+const EditorialContent: React.FC<React.HTMLAttributes<HTMLElement>> = ({
+  children,
+  ...otherProps
+}) => <EditorialContentStyle {...otherProps}>{children}</EditorialContentStyle>
+
+export default EditorialContent
