+++ packages/asc-ui/src/components/Editorial/EditorialContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EditorialContent'
