+++ packages/asc-ui/src/components/Editorial/EditorialHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EditorialHeader'
