+++ packages/asc-ui/src/components/Editorial/EditorialMetaList/index.ts
@@ -1,0 +1,8 @@
+import EditorialMetaListStyle, { EditorialMetaListItemStyle } from './EditorialMetaListStyle'
+
+export { default } from './EditorialMetaList'
+
+export const EditorialMetaListItemStyles = {
+  EditorialMetaListStyle,
+  EditorialMetaListItemStyle,
+}
