+++ packages/asc-ui/src/components/Editorial/EditorialPost/index.ts
@@ -1,0 +1,2 @@
+export { default as EditorialPostStyle } from './EditorialPostStyle'
+export { default as EditorialPost } from './EditorialPost'
