+++ packages/asc-ui/src/components/Editorial/EditorialSidebar/EditorialSidebar.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import EditorialSidebarStyle from './EditorialSidebarStyle'
+
+const EditorialContent: React.FC<React.HTMLAttributes<HTMLElement>> = ({
+  children,
+  ...otherProps
+}) => <EditorialSidebarStyle {...otherProps}>{children}</EditorialSidebarStyle>
+
+export default EditorialContent
