+++ packages/asc-ui/src/components/Editorial/EditorialSidebar/EditorialSidebarStyle.ts
