+++ packages/asc-ui/src/components/Editorial/EditorialSidebar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EditorialSidebar'
