+++ packages/asc-ui/src/components/Editorial/Publication/publication_feed.json
