+++ packages/asc-ui/src/components/Editorial/index.ts
@@ -1,0 +1,20 @@
+import EditorialContentStyle from './EditorialContent/EditorialContentStyle'
+import EditorialHeaderStyle from './EditorialHeader/EditorialHeaderStyle'
+import EditorialSidebarStyle from './EditorialSidebar/EditorialSidebarStyle'
+import EditorialMetaListStyle from './EditorialMetaList/EditorialMetaListStyle'
+import { EditorialPostStyle } from './EditorialPost'
+
+export { EditorialPost } from './EditorialPost'
+export { EditorialBody } from './EditorialBody'
+export { default as EditorialContent } from './EditorialContent'
+export { default as EditorialSidebar } from './EditorialSidebar'
+export { default as EditorialMetaList } from './EditorialMetaList'
+export { default as EditorialHeader } from './EditorialHeader'
+
+export const EditorialStyles = {
+  EditorialPostStyle,
+  EditorialContentStyle,
+  EditorialMetaListStyle,
+  EditorialSidebarStyle,
+  EditorialHeaderStyle,
+}
