--- packages/asc-ui/src/components/Publication/Publication.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import PublicationStyle from './PublicationStyle'
-
-type Props = {}
-
-const Publication: React.FC<Props> = ({ children }) => (
-  <PublicationStyle>{children}</PublicationStyle>
-)
-
-export default Publication
