--- packages/asc-ui/src/components/Publication/PublicationStyle.ts
@@ -1,6 +1,0 @@
-import styled from '@datapunt/asc-core'
-import { BlogPost } from '../Blog/BlogPost'
-
-export type Props = {}
-
-export default styled(BlogPost)<Props>``
