--- packages/asc-ui/src/components/Publication/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Publication'
